requirements.txt
pandas
yfinance
streamlit
phidata
python-dotenv